// lib/components/tic-tac-toe/TicTacToe.enum.ts
export enum GameStatus {
  PLAYING = 'playing',
  WIN = 'win',
  TIE = 'tie'
}

export enum Player {
  X = 'X',
  O = 'O'
}

export enum GameMode {
  CPU = 'cpu',
  HUMAN = 'human'
}

export enum GameLevel {
  EASY = 'easy',
  MEDIUM = 'medium',
  HARD = 'hard'
}

export enum j5TicTacToeEnum {
  tag = "j5-tic-tac-toe",
}